Move modules to more fitting packages
